Rename Xss scanner
